Fixed issue with exception when publishing
In dotnet 6 there are stricter checks if the publish command will try to
copy files with the same name to the same target location.

Because the module has both a linked module.config from the nuget as well
as one that is copied into modules/_protected to support running the
module when developing sites and *.config files are considered as
content by default and therefore are copied during publish this
causes an exception.

Fixed by converting into a zip-file containing the module.config since
that will not be copied from modules/_protected only from the nuget
to the publish target.

Fixes #39