Adds possibility for user to add gadgets to the assets pane
It makes the right side pane into the user type, which allows for
saving customization data to the server about which gadgets to have,
their sizes and so forth.
Sets the allowed gadget types to be cms and content which is the same
as the cms home view.

relates to: #71